+++ packages/base/src/types/index.ts
@@ -1,0 +1,2 @@
+export * from './PolymorphicComponentProps';
+export * from './Domain';