+++ packages/templates/src/components/Example/Example.types.ts
@@ -1,0 +1,3 @@
+export type ExampleProps = {
+  text: string;
+};