+++ packages/templates/src/emotion.d.ts
@@ -1,0 +1,5 @@
+import '@emotion/react';
+import type React from 'react';
+declare module '*.svg' {
+  export const ReactComponent: React.FunctionComponent<React.SVGProps<SVGSVGElement>>;
+}